QA Fields into own package
