+++ app/src/main/java/app/zimly/backup/ui/screens/editor/BucketConfiguration.kt
@@ -28,6 +28,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
+import app.zimly.backup.ui.screens.editor.field.Field
 import app.zimly.backup.ui.theme.containerBackground
 
 @Composable
