+++ app/src/main/java/app/zimly/backup/ui/screens/editor/field/Field.kt
@@ -1,21 +1,20 @@
-package app.zimly.backup.ui.screens.editor
+package app.zimly.backup.ui.screens.editor.field
 
-import android.net.Uri
 import androidx.compose.ui.focus.FocusState
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 
-class UriField(
+class Field(
     private val errorMessage: String = "This field is required.",
-    private val validate: (value: Uri) -> Boolean = { it.path?.isNotEmpty() ?: false },
+    private val validate: (value: String) -> Boolean = { it.isNotEmpty() },
 ) {
     private var touched: Boolean? = null
     private val internal: MutableStateFlow<FieldState> = MutableStateFlow(FieldState())
     val state: StateFlow<FieldState> = internal.asStateFlow()
 
-    fun update(value: Uri) {
+    fun update(value: String) {
         internal.update {
             it.copy(
                 value = value,
@@ -48,5 +47,5 @@ class UriField(
     }
 
 
-    data class FieldState(val value: Uri = Uri.EMPTY, val error: String? = null)
+    data class FieldState(val value: String = "", val error: String? = null)
 }
